# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Costa Rica Vibe API
  description: Directorio de lugares y experiencias de las 7 provincias de Costa Rica, asistente virtual Vibe y panel de administración.
  version: 1.4.2
servers:
  - url: https://localvibe-api-kxga.onrender.com
    description: Prod (Render)
  - url: http://localhost:4000
    description: Local
tags:
  - name: Salud
  - name: Catalogo
  - name: Vibe
    description: Asistente virtual con IA
  - name: Admin
paths:
  /health:
    get:
      tags: [Salud]
      summary: Estado del servicio
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: localvibe-api
  /api/places:
    get:
      tags: [Catalogo]
      summary: Lista de lugares con filtros
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Texto a buscar sobre nombre, descripción, ubicación, vibra o tags.
        - name: category
          in: query
          schema: { type: string }
          description: Categoría exacta (o "todos").
        - name: province
          in: query
          schema: { type: string }
          description: Filtra por provincia (ej. "Limón", "Guanacaste").
      responses:
        "200":
          description: Lista de lugares
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  places:
                    type: array
                    items: { $ref: "#/components/schemas/Place" }
  /api/places/categories:
    get:
      tags: [Catalogo]
      summary: Categorías disponibles en el catálogo
      responses:
        "200":
          description: Lista de categorías con su etiqueta
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        label: { type: string }
  /api/places/{slug}:
    get:
      tags: [Catalogo]
      summary: Detalle de un lugar por slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Lugar encontrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Place" }
        "404":
          $ref: "#/components/responses/NotFound"
  /api/concierge:
    post:
      tags: [Vibe]
      summary: Conversa con Vibe, el asistente virtual
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConciergeRequest" }
      responses:
        "200":
          description: Respuesta del asistente virtual
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConciergeReply" }
        "400":
          description: Mensaje vacío
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
  /api/admin/login:
    post:
      tags: [Admin]
      summary: Inicia sesión de administrador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  format: password
      responses:
        "200":
          description: Token de sesión
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expiresIn:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: Admin no configurado en el entorno
  /api/admin/logout:
    post:
      tags: [Admin]
      summary: Cierra la sesión actual
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sesión cerrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /api/admin/places:
    get:
      tags: [Admin]
      summary: Lista de lugares (panel admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de lugares
          content:
            application/json:
              schema:
                type: object
                properties:
                  places:
                    type: array
                    items: { $ref: "#/components/schemas/Place" }
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Admin]
      summary: Crea un lugar
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlaceInput" }
      responses:
        "201":
          description: Lugar creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  place: { $ref: "#/components/schemas/Place" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/admin/places/{id}:
    put:
      tags: [Admin]
      summary: Actualiza un lugar
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlaceInput" }
      responses:
        "200":
          description: Lugar actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  place: { $ref: "#/components/schemas/Place" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Admin]
      summary: Elimina un lugar
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Lugar eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/export:
    get:
      tags: [Admin]
      summary: Exporta un archivo de notas del directorio
      parameters:
        - name: file
          in: query
          schema: { type: string, default: notas.txt }
      responses:
        "200":
          description: Contenido del archivo
          content:
            text/plain:
              schema: { type: string }
        "500":
          description: No se pudo exportar
  /api/admin/resumen:
    get:
      tags: [Admin]
      summary: Resumen textual del directorio
      responses:
        "200":
          description: Texto del resumen
          content:
            text/plain:
              schema: { type: string }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT-like (token opaco de 8h)
  schemas:
    Place:
      type: object
      required:
        - id
        - slug
        - name
        - category
        - categoryLabel
        - province
        - location
        - vibe
        - description
        - tags
        - price
        - rating
        - hours
        - image
        - featured
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        category:
          type: string
          description: Identificador de la categoría (ej. turismo, cafe, estudio).
        categoryLabel:
          type: string
        province:
          type: string
          description: Provincia de Costa Rica (ej. Limón, Guanacaste).
        location:
          type: string
        vibe:
          type: string
        description:
          type: string
        tags:
          type: array
          items: { type: string }
        price:
          type: string
        rating:
          type: number
          format: float
        hours:
          type: string
        image:
          type: string
          format: uri
        featured:
          type: boolean
    PlaceInput:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
        category:
          type: string
        categoryLabel:
          type: string
        province:
          type: string
        location:
          type: string
        vibe:
          type: string
        description:
          type: string
        tags:
          type: array
          items: { type: string }
        price:
          type: string
        rating:
          type: number
          format: float
        hours:
          type: string
        image:
          type: string
          format: uri
        featured:
          type: boolean
    ConciergeRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
        history:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
    ConciergeReply:
      type: object
      required: [reply, placeIds]
      properties:
        reply:
          type: string
        placeIds:
          type: array
          items: { type: string }
    ErrorBody:
      type: object
      properties:
        error:
          type: string
  responses:
    Unauthorized:
      description: Token faltante o expirado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
    BadRequest:
      description: Datos inválidos
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
    NotFound:
      description: No encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }