# backend/src/routes/places.ts
import { Router } from "express";
import { asyncHandler } from "../lib/asyncHandler";
import { listPlaces } from "../lib/store";
import type { Place } from "../data/places";

const router = Router();

const normalize = (value: string) =>
  value.toLowerCase().normalize("NFD").replace(/[\u0300-\u036f]/g, "");

const placeText = (place: Place) =>
  normalize(
    [place.name, place.description, place.location, place.vibe, place.categoryLabel, place.province ?? "", ...place.tags].join(" ")
  );

router.get(
  "/categories",
  asyncHandler(async (_req, res) => {
    const map = new Map<string, string>();
    for (const place of await listPlaces()) {
      if (!map.has(place.category)) map.set(place.category, place.categoryLabel);
    }
    res.json({ categories: [...map.entries()].map(([id, label]) => ({ id, label })) });
  })
);

router.get(
  "/",
  asyncHandler(async (req, res) => {
    const q = normalize(String(req.query.q ?? "").trim());
    const category = String(req.query.category ?? "").trim().toLowerCase();
    const province = normalize(String(req.query.province ?? "").trim());

    let results = await listPlaces();

    if (category && category !== "todos") {
      results = results.filter((place) => place.category.toLowerCase() === category);
    }

    if (province) {
      results = results.filter((place) => normalize(place.province ?? "").includes(province));
    }

    if (q) {
      results = results.filter((place) => placeText(place).includes(q));
    }

    res.json({ count: results.length, places: results });
  })
);

router.get(
  "/:slug",
  asyncHandler(async (req, res) => {
    const place = (await listPlaces()).find((item) => item.slug === req.params.slug);
    if (!place) {
      res.status(404).json({ error: "Comercio no encontrado" });
      return;
    }
    res.json(place);
  })
);

export default router;